add types to Main Page
