+++ packages/frontend/.eslintrc
@@ -7,7 +7,10 @@
   "extends": ["eslint:recommended", "plugin:react/recommended", "prettier"],
   "rules": {
     "semi": ["error", "always"],
-    "react/react-in-jsx-scope": "off"
+    "react/react-in-jsx-scope": "off",
+    "no-undef": "off",
+    "react/prop-types": "off"
+
   },
   "parserOptions": {
     "sourceType": "module"
