+++ packages/frontend/src/views/MainPage/MainPage.tsx
@@ -11,6 +11,7 @@ import { AccountsList } from '../../components/Accounts/AccountsList';
 import { AddButton } from '../../components/Button/Button';
 import { AccountsContext } from '../../reducers/accounts.reducer';
 import { StyledDate, StyledDateWrap } from './MainPage.styles';
+// @ts-ignore
 import Add from '../../assets/add.png';
 
 const MainPage = () => {
@@ -22,7 +23,7 @@ const MainPage = () => {
   const [expenses, setExpenses] = useState(0);
   const [incomes, setIncomes] = useState(0);
   const [balance, setBalance] = useState(0);
-  const token = JSON.parse(localStorage.getItem('user'))?.token;
+  const token = JSON.parse((localStorage.getItem('user')) as string)?.token;
 
   useEffect(() => {
     getInitialList();
@@ -59,7 +60,8 @@ const MainPage = () => {
   };
 
   const { accountsState, limitsState } = useContext(AccountsContext);
-  const [limits, limitsDispatch] = limitsState;
+  // TODO REMOVE ANY
+  const [limits, limitsDispatch] = limitsState as any;
   const onSetShowAlert = () => limitsDispatch({ type: 'closeLastLimit' });
 
   const [list, dispatch] = accountsState;
